+++ packages/core/styles/src/packages/atomic/link/link.style.css.ts
@@ -16,6 +16,10 @@ const base = vanillaStyle({
   boxSizing: "border-box",
 
   transition: `all ${varsThemeBase.motion.speed.fast} ease`,
+
+  ":focus": {
+    boxShadow: varsThemeBase.utils.focus,
+  },
 });
 
 export const size = styleVariants({
@@ -51,9 +55,6 @@ export const appearance = styleVariants({
       ":active": {
         color: varsThemeBase.colors.primary.interactivePressed,
       },
-      ":focus": {
-        outline: `${varsThemeBase.colors.primary.interactivePressed} auto 1px`,
-      },
     },
   ],
   danger: [
@@ -66,9 +67,6 @@ export const appearance = styleVariants({
       ":active": {
         color: varsThemeBase.colors.danger.interactivePressed,
       },
-      ":focus": {
-        outline: `${varsThemeBase.colors.danger.interactivePressed} auto 1px`,
-      },
     },
   ],
   neutral: [
@@ -80,9 +78,6 @@ export const appearance = styleVariants({
       },
       ":active": {
         color: varsThemeBase.colors.neutral.textHigh,
-      },
-      ":focus": {
-        outline: `${varsThemeBase.colors.neutral.textHigh} auto 1px`,
       },
     },
   ],