+++ packages/react/src/atomic/Button/src/button.types.ts
@@ -9,7 +9,7 @@ export interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
   /** Text to be displayed on chip */
   children: ReactNode;
   /** Button appearance */
-  appearance?: "primary" | "danger" | "neutral";
+  appearance?: "primary" | "danger" | "neutral" | "transparent";
   /** Set whether button state is disabled */
   disabled?: boolean;
 }